Move iptables code to pkg/iptables
