+++ pkg/iptables/iptables.go
@@ -1,4 +1,4 @@
-package main
+package iptables
 
 import (
 	"fmt"
